Make character engine (-Tascii, -Tpdf, -Tps) ready for Unicode: make buffer consist of type "int".  This will take more work (especially in encode and friends), but this is a strong start.  This commit also consists of some harmless lint fixes.

+++ mandoc.c
@@ -704,7 +704,7 @@ mandoc_strntou(const char *p, size_t sz, int base)
 		return(-1);
 
 	memcpy(buf, p, sz);
-	buf[sz] = '\0';
+	buf[(int)sz] = '\0';
 
 	errno = 0;
 	v = strtol(buf, &ep, base);
